requirements.txt
pandas==2.0.2
langchain==0.0.271
python-dotenv==1.0.0